Create DeployAnL1_Codespace.md
We are going to be walking through how to deploy our own test blockchain with 5 nodes, with our own token. We'll also showcase how to deploy a simple smart contract on this newly created blockchain. We will be using the Avalanche Start Kit repo, Docker, and Github Codespaces to run this project -- meaning you don't have to install anything :)

Signed-off-by: Usman Asim <97273722+usmaneth@users.noreply.github.com>